split inbound worker handlers
