+++ crates/apps/reticulumd/src/bin/reticulumd/inbound_propagation.rs
@@ -1,4 +1,8 @@
 use super::*;
+use lxmf::inbound_decode::InboundPayloadMode;
+use reticulum_daemon::inbound_delivery::{
+    annotate_inbound_record_stamp_status, decode_inbound_payload, evaluate_inbound_stamp_policy,
+};
 use x25519_dalek::PublicKey;
 
 pub(super) fn is_lxmf_propagation_destination(
